+++ sample/src/main/kotlin/com/thefuntasty/donutsample/tools/extensions/AppCompatActivityExtensions.kt
@@ -1,0 +1,6 @@
+package com.thefuntasty.donutsample.tools.extensions
+
+import androidx.appcompat.app.AppCompatActivity
+import androidx.core.content.ContextCompat
+
+fun AppCompatActivity.getColorCompat(id: Int) = ContextCompat.getColor(this, id)
