+++ sample/src/main/kotlin/com/thefuntasty/donutsample/tools/extensions/SeekBarExtensions.kt
@@ -1,0 +1,17 @@
+package com.thefuntasty.donutsample.tools.extensions
+
+import android.widget.SeekBar
+
+fun SeekBar.doOnProgressChange(f: (progress: Int) -> Unit) {
+    setOnSeekBarChangeListener(object : SeekBar.OnSeekBarChangeListener {
+        override fun onProgressChanged(seekBar: SeekBar?, progress: Int, fromUser: Boolean) {
+            if (fromUser) {
+                f(progress)
+            }
+        }
+
+        override fun onStartTrackingTouch(seekBar: SeekBar?) = Unit
+
+        override fun onStopTrackingTouch(seekBar: SeekBar?) = Unit
+    })
+}
